add unit tests for total, fix bug
categories that are allowed when not graphing by individual member
(i.e. categories that involve percentages of the whole and would always be 100% when graphing total),
were not added to the output dict when not graphing by individual member.
now that more code has been moved into get_results, those categories end up getting filled in
anyway and need to be initialized